+++ packages/logger/src/index.ts
@@ -3,4 +3,4 @@
 // This file is licensed under the MIT License.
 // License text available at https://opensource.org/licenses/MIT
 
-export {Logger, LoggingOptions} from './logging';
+export {Logger, LoggingOptions} from './logger';
