+++ packages/logger/test/integration/smoke.test.ts
