+++ packages/logger/tsconfig.build.json
