+++ packages/openapi-spec-builder/index.js
@@ -1,9 +1,12 @@
-// Copyright IBM Corp. 2017. All Rights Reserved.
-// Node module: @loopback/testlab
+"use strict";
+// Copyright IBM Corp. 2013,2017. All Rights Reserved.
+// Node module: @loopback/openapi-spec-builder
 // This file is licensed under the MIT License.
 // License text available at https://opensource.org/licenses/MIT
-
-const nodeMajorVersion = +process.versions.node.split('.')[0];
-module.exports = nodeMajorVersion >= 7 ?
-  require('./lib') :
-  require('./lib6');
+function __export(m) {
+    for (var p in m) if (!exports.hasOwnProperty(p)) exports[p] = m[p];
+}
+Object.defineProperty(exports, "__esModule", { value: true });
+// NOTE(bajtos) This file is used by VSCode/TypeScriptServer at dev time only
+__export(require("./src"));
+//# sourceMappingURL=index.js.map
