+++ packages/openapi-spec/index.d.ts
@@ -1,6 +1,1 @@
-// Copyright IBM Corp. 2017. All Rights Reserved.
-// Node module: @loopback/openapi-spec
-// This file is licensed under the MIT License.
-// License text available at https://opensource.org/licenses/MIT
-
-export * from './lib';
+export * from './src';
